apps/ai-backend/pyproject.toml
[tool.poetry]
name = "fitos-ai-backend"
version = "0.1.0"
description = "LangGraph-based AI coaching backend for FitOS"
authors = ["FitOS Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
langgraph = "^1.0.0"
langgraph-checkpoint = "^2.0.0"
langchain = "^0.3.9"
langchain-openai = "^0.2.9"
langchain-anthropic = "^0.2.5"
langsmith = "^0.2.0"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
httpx = "^0.28.0"
python-multipart = "^0.0.17"
pillow = "^11.0.0"
numpy = "^2.2.1"
supabase = "^2.10.0"
websockets = "^14.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
mcp = "^1.0.0"
anthropic = "^0.39.0"
deepgram-sdk = "^3.7.0"
elevenlabs = "^1.8.0"
pyaudio = "^0.2.14"
sounddevice = "^0.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"
ruff = "^0.8.4"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
